Cleaning and adding MongoDB support

+++ src/main/java/com/example/repository/AuthorRepository.java
@@ -3,8 +3,7 @@ package com.example.repository;
 import com.example.domain.Author;
 import org.springframework.data.repository.CrudRepository;
 
-/**
- * Created by adria on 06.05.2017.
- */
-public interface AuthorRepository extends CrudRepository<Author, Long>{
+import java.math.BigInteger;
+
+public interface AuthorRepository extends CrudRepository<Author, BigInteger>{
 }
